algorithms/dynamic_programming.py
def count_palindromes_dynamic(s):
    """
    Count the number of palindromes in a given string using dynamic programming method.
    
    Args:
    s (str): The input string
    
    Returns:
    int: Number of palindromes found in the input string
    """
    n = len(s)                                  # O(1)
    dp = [[False] * n for _ in range(n)]        # O(n^2)
    count = 0                                   # O(1)
    for i in range(n):                          # O(n)
        dp[i][i] = True                         # O(1)
        count += 1                              # O(1)
    for i in range(n-1):                        # O(n)
        if s[i] == s[i+1]:                      # O(1)
            dp[i][i+1] = True                   # O(1)
            count += 1                          # O(1)
    for length in range(3, n+1):                # O(n)
        for i in range(n-length+1):             # O(n)
            j = i + length - 1                  # O(1)
            if s[i] == s[j] and dp[i+1][j-1]:   # O(1)
                dp[i][j] = True                 # O(1)
                count += 1                      # O(1)
    return count                                # O(1)

# Complexity: O(n^2) + O(n) + O(n) + O(n^2) ≈ O(2n^2) + O(2n) ≈ O(n^2)